chore: extract core/game_rng from utils/game_utils
