+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Changed
 
-- Extract `core/game_log` and `core/game_strings` from `utils/game_utils`
+- Extract `core/game_log`, `core/game_strings`, `core/game_rng` from `utils/game_utils`
 - Build `helapordo_core.o` separately from `core`
 - Move `core/game_animations` to `animations/game_animations`
 - Move `setSprite` functions `core/sprites`
