+++ scripts/save2l/save2l.c
@@ -23,7 +23,6 @@ int GS_AUTOSAVE_ON = 1;
 int G_USE_DEFAULT_BACKGROUND = 0;
 int G_USE_VIM_DIRECTIONAL_KEYS = 0;
 int G_USE_WASD_DIRECTIONAL_KEYS = 0;
-int64_t G_RNG_ADVANCEMENTS = 0;
 int G_SEEDED_RUN_ON = 0;
 char *G_SEEDED_RUN_ARG = NULL;
 Gamemode GAMEMODE = Rogue;
